backend/pyproject.toml
[tool.poetry]
name = "AISCL-backend"
version = "1.0.0"
description = "AISCL Backend API"
authors = ["AISCL Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
motor = "^3.3.0"
beanie = "^1.23.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
redis = "^5.0.0"
minio = "^7.2.0"
langchain = "^1.0.0"
langgraph = "^1.0.0"
langchain-openai = "^1.0.0"
openai = "^1.30.0"
tiktoken = "^0.8.0"
beautifulsoup4 = "^4.12.0"
playwright = {extras = ["asyncio"], version = "^1.40.0"}
prometheus-fastapi-instrumentator = "^6.1.0"
slowapi = "^0.1.9"
sse-starlette = "^1.8.0"

aiofiles = "^23.2.0"
python-socketio = "^5.10.0"
ypy-websocket = "^0.12.0"
email-validator = "^2.3.0"
bcrypt = "4.0.1"
langchain-community = "^0.4.1"
sentence-transformers = "^5.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.26.0"
black = "^23.12.0"
isort = "^5.13.0"
pylint = "^3.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pylint.messages_control]
max-line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

